requirements.txt
influxdb==5.2.2
paho-mqtt==1.4.0
YURL==1.0.0
